qcacld-3.0: Don't use hdd_assemble_rate_code API to set 11be rate
hdd_assemble_rate_code uses WMI_ASSEMBLE_RATECODE_V1 to configure rate
to Fw. But this macro is not as per expectation by FW.

To address this, Don't use hdd_assemble_rate_code API and read the rate
code from user as per FW expectation and configure it to the FW.

To set 11be rate, write 2 bytes value to the sysfs entry as below.
In this byte-1 should be 11be preamble id i.e. 5, in byte-2 MSB 3-bits
represents NSS and LSB 5-bits represents MCS. For e.g. the below
command will configure the 11be rate to MCS 7 and NSS 2.

echo 0x0547 > /sys/class/net/wlan1/11be_rate

CRs-Fixed: 3004042

+++ core/hdd/inc/wlan_hdd_eht.h
@@ -97,7 +97,7 @@ void wlan_hdd_get_mlo_link_id(struct hdd_beacon_data *beacon,
  *
  * Return: 0 on success, negative errno on failure
  */
-int hdd_set_11be_rate_code(struct hdd_adapter *adapter, uint8_t rate_code);
+int hdd_set_11be_rate_code(struct hdd_adapter *adapter, uint16_t rate_code);
 
 /**
  * hdd_sysfs_11be_rate_create() - Create sysfs entry to configure 11be rate
@@ -132,7 +132,7 @@ void hdd_update_wiphy_eht_cap(struct hdd_context *hdd_ctx)
 }
 
 static inline int
-hdd_set_11be_rate_code(struct hdd_adapter *adapter, uint8_t rate_code)
+hdd_set_11be_rate_code(struct hdd_adapter *adapter, uint16_t rate_code)
 {
 	return 0;
 }
